+++ react/components/ProductImages/components/Carousel/index.js
@@ -1,100 +1,70 @@
 /* eslint-disable react/prop-types */
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import debounce from 'debounce'
 import classNames from 'classnames'
 import { path, equals } from 'ramda'
-import ReactResizeDetector from 'react-resize-detector'
 import { IconCaret } from 'vtex.store-icons'
-import { withCssHandles, applyModifiers } from 'vtex.css-handles'
+import { withCssHandles } from 'vtex.css-handles'
+import SwiperCore, { Thumbs, Navigation, Pagination } from 'swiper'
+import { Swiper, SwiperSlide } from 'swiper/react'
 
 import Video, { getThumbUrl } from '../Video'
 import ProductImage from '../ProductImage'
-import styles from '../../styles.css'
-import './global.css'
 import ThumbnailSwiper from './ThumbnailSwiper'
 import {
   THUMBS_ORIENTATION,
   THUMBS_POSITION_HORIZONTAL,
 } from '../../utils/enums'
+import styles from './swiper.scoped.css'
 
-/**
- * ReactIdSwiper cannot be SSRendered, so this is a fake swiper that copies some of its classes and HTML layout and render only the first image of the children array.
- */
-const FakeSwiper = ({
-  children,
-  containerClass,
-  direction = THUMBS_ORIENTATION.HORIZONTAL,
-}) => {
-  const swiperContainerDirection =
-    direction === THUMBS_ORIENTATION.HORIZONTAL
-      ? 'swiper-container-horizontal'
-      : direction === THUMBS_ORIENTATION.VERTICAL
-      ? 'swiper-container-vertical'
-      : ''
+import './swiper.global.css'
+import './overrides.global.css'
 
-  const childrenArray = React.Children.toArray(children)
+const CARET_ICON_SIZE = 24
+const CARET_CLASSNAME =
+  'pv8 absolute top-50 translate--50y z-2 pointer c-action-primary'
 
-  if (childrenArray.length === 0) {
-    return null
-  }
-
-  const [child] = childrenArray
-  const childClass = path(['props', 'className'], child)
-  const newChildClass = childClass
-    ? `${childClass} swiper-slide-active`
-    : childClass
-
-  return (
-    <div
-      className={`${containerClass} swiper-container-initialized ${swiperContainerDirection}`}
-    >
-      <div className="swiper-wrapper">
-        {React.cloneElement(child, {
-          className: newChildClass,
-        })}
-      </div>
-    </div>
-  )
-}
-
-/** Swiper and its modules are imported using require to avoid breaking SSR */
-const Swiper = window.navigator
-  ? require('react-id-swiper/lib/ReactIdSwiper').default
-  : FakeSwiper
-// eslint-disable-next-line padding-line-between-statements
-const SwiperModules = window.navigator ? require('swiper/dist/js/swiper') : {}
+// install Swiper's Thumbs component
+SwiperCore.use([Thumbs, Navigation, Pagination])
 
 const CSS_HANDLES = [
   'carouselContainer',
   'productImagesThumbsSwiperContainer',
+  'productImagesThumbActive',
   'productImagesGallerySwiperContainer',
   'productImagesGallerySlide',
   'swiperCaret',
   'swiperCaretNext',
   'swiperCaretPrev',
-  'productImagesThumbCaret',
-  'swiperBullet',
 ]
 
 const initialState = {
-  loaded: [],
   thumbUrl: [],
   alt: [],
-  thumbsLoaded: false,
   activeIndex: 0,
 }
 
 class Carousel extends Component {
-  thumbSwiper = null
-  gallerySwiper = null
-  state = initialState
+  state = {
+    ...initialState,
+    thumbSwiper: null,
+    gallerySwiper: null,
+  }
 
-  async setInitialVariablesState() {
+  isVideo = []
+
+  get hasGallerySwiper() {
+    return Boolean(this.state.gallerySwiper)
+  }
+
+  get hasThumbSwiper() {
+    return Boolean(this.state.thumbSwiper)
+  }
+
+  setInitialVariablesState() {
     const slides = this.props.slides || []
 
     this.isVideo = []
-    this.thumbLoadCount = 0
 
     slides.forEach(async (slide, i) => {
       if (slide.type === 'video') {
@@ -102,107 +72,71 @@ class Carousel extends Component {
 
         this.isVideo[i] = true
         this.setVideoThumb(i)(thumbUrl)
-        this.thumbLoadFinish()
       } else {
-        this.getThumb(slide.thumbUrl)
+        // Image object doesn't exist when it's being rendered in the server side
+        if (!window.navigator) {
+          return
+        }
+
+        const image = new Image()
+
+        image.src = slide.thumbUrl
       }
     })
   }
 
-  updateSwiperSize = debounce(() => {
-    if (this.thumbSwiper) {
-      this.thumbSwiper.update()
-    }
-
-    if (this.gallerySwiper) {
-      this.gallerySwiper.update()
-    }
-  }, 500)
-
-  thumbLoadFinish = () => {
-    this.thumbLoadCount++
-    if (
-      !this.props.slides ||
-      this.thumbLoadCount === this.props.slides.length
-    ) {
-      this.setState({ thumbsLoaded: true })
-    }
-  }
-
-  getThumb = thumbUrl => {
-    if (!window.navigator) return // Image object doesn't exist when it's being rendered in the server side
-    const image = new Image()
-
-    image.onload = () => {
-      this.thumbLoadFinish()
-    }
-
-    image.onerror = () => {
-      this.thumbLoadFinish()
-    }
-
-    image.src = thumbUrl
-  }
-
-  handleResize = () => {
-    this.updateSwiperSize()
-  }
-
   componentDidMount() {
-    window.addEventListener('resize', this.handleResize)
     this.setInitialVariablesState()
   }
 
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.handleResize)
-
-    this.updateSwiperSize.clear()
-  }
-
   componentDidUpdate(prevProps) {
-    const { loaded, activeIndex } = this.state
+    const { activeIndex } = this.state
     const { isVideo } = this
 
     if (!equals(prevProps.slides, this.props.slides)) {
       this.setInitialVariablesState()
-      this.setState(initialState)
-      if (this.props.slides && this.props.slides.length > 1) {
-        this.gallerySwiper && this.gallerySwiper.slideTo(0)
-        this.thumbSwiper && this.thumbSwiper.slideTo(0)
-      }
+
+      const newInitialState = { ...initialState }
+
+      if (!this.props.slides) return
+
+      this.setState(newInitialState)
 
       return
     }
 
-    const paginationElement = path(['pagination', 'el'], this.gallerySwiper)
+    const paginationElement = path(
+      ['swiper', 'pagination', 'el'],
+      this.state.gallerySwiper
+    )
 
-    if (paginationElement) paginationElement.hidden = isVideo[activeIndex]
-
-    const gallerySwiperZoom = path(['zoom'], this.gallerySwiper)
-
-    if (gallerySwiperZoom) {
-      loaded[activeIndex]
-        ? gallerySwiperZoom.enable()
-        : gallerySwiperZoom.disable()
+    if (paginationElement) {
+      paginationElement.hidden = isVideo[activeIndex]
     }
   }
 
-  onSlideChange = () => {
-    const activeIndex = path(['activeIndex'], this.gallerySwiper)
+  handleSlideChange = () => {
+    this.setState(prevState => {
+      if (!this.hasGallerySwiper) {
+        return
+      }
 
-    this.setState({ activeIndex, sliderChanged: true })
+      const { activeIndex } = prevState.gallerySwiper
+
+      return { activeIndex, sliderChanged: true }
+    })
   }
 
   setVideoThumb = i => (url, title) => {
-    // eslint-disable-next-line react/no-access-state-in-setstate
-    const thumbUrl = { ...this.state.thumbUrl }
-    // eslint-disable-next-line react/no-access-state-in-setstate
-    const alt = { ...this.state.alt }
+    this.setState(prevState => {
+      const thumbUrl = { ...prevState.thumbUrl }
+      const alt = { ...prevState.alt }
 
-    thumbUrl[i] = url
-    alt[i] = title
+      thumbUrl[i] = url
+      alt[i] = title
 
-    this.setState({ thumbUrl, alt })
+      return { thumbUrl, alt }
+    })
   }
 
   renderSlide = (slide, i) => {
@@ -250,176 +184,41 @@ class Carousel extends Component {
   }
 
   get galleryParams() {
-    const {
-      cssHandles,
-      slides = [],
-      showPaginationDots = true,
-      showNavigationArrows = true,
-    } = this.props
+    const { cssHandles, slides = [], showPaginationDots = true } = this.props
 
-    const iconSize = 24
-    const caretClassName =
-      'pv8 absolute top-50 translate--50y z-2 pointer c-action-primary'
+    const params = {}
 
-    return {
-      modules: [SwiperModules.Pagination, SwiperModules.Navigation],
-      containerClass: `swiper-container ${cssHandles.productImagesGallerySwiperContainer}`,
-      ...(slides.length > 1 &&
-        showPaginationDots && {
-          pagination: {
-            el: '.swiper-pagination',
-            clickable: true,
-            // custom props from vtex/swiper
-            bulletSelector: `.swiper-pagination-bullet`,
-            bulletClass: `swiper-pagination-bullet ${cssHandles.swiperBullet}`,
-            bulletActiveClass: `c-action-primary swiper-pagination-bullet-active ${applyModifiers(
-              cssHandles.swiperBullet,
-              'active'
-            )}`,
-          },
-        }),
-      ...(slides.length > 1 && {
-        navigation: {
-          prevEl: '.swiper-caret-prev',
-          nextEl: '.swiper-caret-next',
-          disabledClass: `c-disabled ${styles.carouselCursorDefault}`,
+    if (slides.length > 1 && showPaginationDots) {
+      params.pagination = {
+        el: `.${styles['swiper-pagination']}`,
+        clickable: true,
+        clickableClass: styles.swiperPaginationClickable,
+        bulletClass: styles.swiperBullet,
+        bulletActiveClass: styles['swiperBullet--active'],
+        renderBullet(_index, className) {
+          return `<span class="${className} c-action-primary"></span>`
         },
-      }),
-      thumbs: {
-        swiper: this.thumbSwiper,
-      },
-      threshold: 10,
-      resistanceRatio: slides.length > 1 ? 0.85 : 0,
-      ...(showNavigationArrows && {
-        renderNextButton: () => (
-          <span
-            className={`swiper-caret-next pl7 pr2 right-0 ${caretClassName} ${cssHandles.swiperCaret} ${cssHandles.swiperCaretNext}`}
-          >
-            <IconCaret
-              orientation="right"
-              size={iconSize}
-              className={styles.carouselIconCaretRight}
-            />
-          </span>
-        ),
-        renderPrevButton: () => (
-          <span
-            className={`swiper-caret-prev pr7 pl2 left-0 ${caretClassName} ${cssHandles.swiperCaret} ${cssHandles.swiperCaretPrev}`}
-          >
-            <IconCaret
-              orientation="left"
-              size={iconSize}
-              className={styles.carouselIconCaretLeft}
-            />
-          </span>
-        ),
-      }),
-      on: {
-        slideChange: this.onSlideChange,
-      },
-      getSwiper: swiper => {
-        if (this.gallerySwiper !== swiper) {
-          this.gallerySwiper = swiper
-        }
-      },
+      }
     }
-  }
 
-  get thumbnailsParams() {
-    const {
-      displayThumbnailsArrows,
-      thumbnailsOrientation,
-      cssHandles,
-    } = this.props
-
-    const isThumbsVertical =
-      thumbnailsOrientation === THUMBS_ORIENTATION.VERTICAL
-
-    const caretSize = 24
-    const caretClassName = `${cssHandles.productImagesThumbCaret} absolute z-2 pointer c-action-primary flex pv2`
-    const caretStyle = { transition: 'opacity 200ms' }
-
-    return {
-      modules: [SwiperModules.Navigation],
-      ...(displayThumbnailsArrows && {
-        navigation: {
-          prevEl: '.swiper-thumbnails-caret-prev',
-          nextEl: '.swiper-thumbnails-caret-next',
-          disabledClass: `c-disabled o-0 pointer-events-none ${styles.carouselCursorDefault}`,
-          hiddenClass: 'dn',
-        },
-        renderNextButton: () => {
-          const classes = classNames(
-            'swiper-thumbnails-caret-next',
-            caretClassName,
-            {
-              [`bottom-0 pt7 left-0 justify-center w-100 ${styles.gradientBaseBottom}`]: isThumbsVertical,
-              [`right-0 top-0 items-center h-100 pl6 ${styles.gradientBaseRight}`]: !isThumbsVertical,
-            }
-          )
-
-          return (
-            <span className={classes} style={caretStyle}>
-              <IconCaret
-                orientation={isThumbsVertical ? 'down' : 'right'}
-                size={caretSize}
-              />
-            </span>
-          )
-        },
-        renderPrevButton: () => {
-          const classes = classNames(
-            'swiper-thumbnails-caret-prev top-0 left-0',
-            caretClassName,
-            {
-              [`pb7 justify-center w-100 ${styles.gradientBaseTop}`]: isThumbsVertical,
-              [`items-center h-100 pr6 ${styles.gradientBaseLeft}`]: !isThumbsVertical,
-            }
-          )
-
-          return (
-            <span className={classes} style={caretStyle}>
-              <IconCaret
-                orientation={isThumbsVertical ? 'up' : 'left'}
-                size={caretSize}
-              />
-            </span>
-          )
-        },
-      }),
-      observer: true,
-      containerClass: `swiper-container h-100 ${cssHandles.productImagesThumbsSwiperContainer}`,
-      watchSlidesVisibility: true,
-      watchSlidesProgress: true,
-      freeMode: false,
-      // It seems crazy but this is just
-      // an workaround to make it work the thumbnails with the carousel
-      slideActiveClass: 'undefined',
-      slideNextClass: 'undefined',
-      slidePrevClass: 'undefined',
-      direction: thumbnailsOrientation,
-      slidesPerView: 'auto',
-      touchRatio: 1,
-      mousewheel: false,
-      preloadImages: true,
-      shouldSwiperUpdate: true,
-      zoom: false,
-      threshold: 8,
-      /* Slides are grouped when thumbnails arrows are enabled
-       * so that clicking on next/prev will scroll more than
-       * one thumbnail */
-      slidesPerGroup: displayThumbnailsArrows ? 4 : 1,
-      getSwiper: swiper => {
-        if (this.thumbSwiper !== swiper) {
-          this.thumbSwiper = swiper
-        }
-      },
+    if (slides.length > 1) {
+      params.navigation = {
+        prevEl: '.swiper-caret-prev',
+        nextEl: '.swiper-caret-next',
+        disabledClass: `c-disabled ${styles.carouselCursorDefault}`,
+      }
     }
+
+    params.thumbs = {
+      swiper: this.state.thumbSwiper,
+      multipleActiveThumbs: false,
+      slideThumbActiveClass: cssHandles.productImagesThumbActive,
+    }
+
+    return params
   }
 
   render() {
-    const { thumbsLoaded, activeIndex } = this.state
-
     const {
       position,
       cssHandles,
@@ -428,6 +227,9 @@ class Carousel extends Component {
       thumbnailAspectRatio,
       thumbnailsOrientation,
       zoomProps: { zoomType },
+      showPaginationDots = true,
+      showNavigationArrows = true,
+      displayThumbnailsArrows = false,
     } = this.props
 
     const isThumbsVertical =
@@ -455,19 +257,16 @@ class Carousel extends Component {
       }
     )
 
-    const thumbnailSwiper = thumbsLoaded && hasThumbs && (
+    const thumbnailSwiper = (
       <ThumbnailSwiper
+        onSwiper={instance => this.setState({ thumbSwiper: instance })}
         isThumbsVertical={isThumbsVertical}
-        slides={slides}
-        activeIndex={activeIndex}
-        swiperParams={this.thumbnailsParams}
-        thumbUrls={this.state.thumbUrl}
-        position={position}
-        onThumbClick={index =>
-          this.gallerySwiper && this.gallerySwiper.slideTo(index)
-        }
         thumbnailAspectRatio={thumbnailAspectRatio}
         thumbnailMaxHeight={thumbnailMaxHeight}
+        thumbUrls={this.state.thumbUrl}
+        displayThumbnailsArrows={displayThumbnailsArrows}
+        slides={slides}
+        position={position}
       />
     )
 
@@ -486,25 +285,64 @@ class Carousel extends Component {
       }
     )
 
-    const SliderComponent = slides.length === 1 ? FakeSwiper : Swiper
-
     return (
       <div className={containerClasses} aria-hidden="true">
         {isThumbsVertical && thumbnailSwiper}
+
         <div className={imageClasses}>
-          {/* eslint-disable-next-line react/jsx-handler-names */}
-          <ReactResizeDetector handleHeight onResize={this.updateSwiperSize}>
-            <SliderComponent {...this.galleryParams} shouldSwiperUpdate>
-              {slides.map((slide, i) => (
-                <div
-                  key={i}
-                  className={`${cssHandles.productImagesGallerySlide} swiper-slide center-all`}
-                >
-                  {this.renderSlide(slide, i)}
-                </div>
-              ))}
-            </SliderComponent>
-          </ReactResizeDetector>
+          <Swiper
+            onSwiper={instance => this.setState({ gallerySwiper: instance })}
+            className={cssHandles.productImagesGallerySwiperContainer}
+            threshold={10}
+            resistanceRatio={slides.length > 1 ? 0.85 : 0}
+            onSlideChange={this.handleSlideChange}
+            updateOnWindowResize
+            {...this.galleryParams}
+          >
+            {slides.map((slide, i) => (
+              <SwiperSlide
+                key={`slider-${i}`}
+                className={`${cssHandles.productImagesGallerySlide} swiper-slide center-all`}
+              >
+                {this.renderSlide(slide, i)}
+              </SwiperSlide>
+            ))}
+
+            <div
+              key="pagination"
+              className={classNames(styles['swiper-pagination'], {
+                dn: slides.length === 1 || !showPaginationDots,
+              })}
+            />
+
+            <div
+              className={classNames({
+                dn: slides.length === 1 || !showNavigationArrows,
+              })}
+            >
+              <span
+                key="caret-next"
+                className={`swiper-caret-next pl7 pr2 right-0 ${CARET_CLASSNAME} ${cssHandles.swiperCaret} ${cssHandles.swiperCaretNext}`}
+              >
+                <IconCaret
+                  orientation="right"
+                  size={CARET_ICON_SIZE}
+                  className={styles.carouselIconCaretRight}
+                />
+              </span>
+              <span
+                key="caret-prev"
+                className={`swiper-caret-prev pr7 pl2 left-0 ${CARET_CLASSNAME} ${cssHandles.swiperCaret} ${cssHandles.swiperCaretPrev}`}
+              >
+                <IconCaret
+                  orientation="left"
+                  size={CARET_ICON_SIZE}
+                  className={styles.carouselIconCaretLeft}
+                />
+              </span>
+            </div>
+          </Swiper>
+
           {!isThumbsVertical && thumbnailSwiper}
         </div>
       </div>