+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [`ProductImages`] Updated Swiper from `v4` to `v6`.
 
 ## [3.125.0] - 2020-08-28
 ### Changed